--- developer/src/common/web/utils/index.ts
@@ -1,5 +1,0 @@
-export { validateMITLicense } from './src/validate-mit-license.js';
-export { KeymanSentry, SentryNodeOptions } from './src/KeymanSentry.js';
-export { getOption, loadOptions, clearOptions } from './src/options.js';
-export { escapeMarkdownChar } from './src/markdown.js';
-export { KeymanUrls } from './src/keyman-urls.js';
