+++ developer/src/common/web/utils/src/index.ts
@@ -1,0 +1,6 @@
+export { validateMITLicense } from './utils/validate-mit-license.js';
+export { KeymanSentry, SentryNodeOptions } from './utils/KeymanSentry.js';
+export { getOption, loadOptions, clearOptions } from './utils/options.js';
+export { escapeMarkdownChar } from './utils/markdown.js';
+export { KeymanUrls } from './utils/keyman-urls.js';
+
