+++ developer/src/common/web/utils/src/utils/validate-mit-license.ts
