+++ developer/src/common/web/utils/tsconfig.json
@@ -7,7 +7,6 @@
       "baseUrl": ".",
     },
   "include": [
-    "index.ts",
     "src/**/*.ts",
   ],
 }
